requirements.txt
flask
pymongo
beautifulsoup4==4.9.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
fake-useragent==0.1.11
future==0.18.2
google-auth==1.14.2
google-auth-oauthlib==0.4.1
gspread==3.6.0
idna==2.9
lxml==4.5.0
myawis==0.2.4
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
pprint==0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-awis==1.2.1
python-dateutil==2.8.1
pytrends==4.7.2
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
xmltodict==0.12.0
xmlutil==0.0.3
xmlutils==1.4
